requirements.txt
opencv-python==4.6.0.66
matplotlib==3.5.2
pandas==1.5.0
fiftyone==0.17.2
albumentations==1.3.0
tensorflow_probability
python-box
kaggle
wandb
tqdm
nvitop
pydot
click
wget
keras-unet-collection